+++ klibs/KLUserInterface.py
@@ -148,6 +148,54 @@ def mouse_clicked(button=None, released=False, within=None, queue=None):
 	return clicked
 
 
+def get_clicks(button=None, released=False, queue=None):
+	"""Gets the (x, y) pixel coordinates of any mouse clicks within the event queue.
+	
+	If a button is specified, this function will only return the pixel coordinates for
+	clicks of that button. Otherwise, this function will return click coordinates for
+	all buttons. Valid button names are ``'left'``, ``'right'``, and ``'middle'``.
+	
+ 	If an event queue is not provided, this function will fetch and clear the current
+	contents of the input event queue.
+
+	Args:
+		button (str, optional): The name of the button to check for clicks. Defaults to
+			``None`` (returns clicks from all buttons).
+		released (bool, optional): If True, this function will return the coordinates
+			for mouse button release events instead of mouse button click events.
+			Defaults to False.
+		queue (:obj:`List` of :obj:`sdl2.SDL_Event`, optional): A list of events to
+			checkfor valid mouse button events.
+
+	Returns:
+		:obj:`List`: A list containing the (x, y) coordinates of any matching mouse
+		click events.
+
+	Raises:
+		ValueError: If an invalid button name is provided.
+
+	"""
+	buttons = {
+		'left': SDL_BUTTON_LEFT, 'right': SDL_BUTTON_RIGHT, 'middle': SDL_BUTTON_MIDDLE,
+	}
+	if button:
+		if not button in buttons.keys():
+			raise ValueError("'{0}' is not a valid mouse button name.".format(button))
+		button = buttons[button]
+
+	clicks = []
+	if queue == None:
+		queue = pump(True)
+	for e in queue:
+		if e.type == SDL_KEYDOWN:
+			ui_request(e.key.keysym)
+		elif e.type == (SDL_MOUSEBUTTONUP if released else SDL_MOUSEBUTTONDOWN):
+			if not button or e.button.button == button:
+				clicks.append((e.button.x, e.button.y))
+
+	return clicks
+
+
 def show_cursor():
 	"""Shows the mouse cursor if it is currently hidden.